Streaming: Refactor move database and redis logic into separate files (#31567)
